Rework to use File::Temp::tempdir for all builds.

+++ orb-install.pl
@@ -21,10 +21,12 @@ use strict;
 use warnings;
 
 use Getopt::Long;
-use Env qw(HOME LDFLAGS CPPFLAGS orb_cache orb_base);
+use Env
+  qw(HOME LDFLAGS CPPFLAGS TMPDIR orb_base orb_cache orb_logs orb_ruby_base orb_perl_base orb_python_base);
 use POSIX qw(getcwd);
 use File::Path qw(make_path);
 use File::Basename qw(dirname basename);
+use File::Temp qw(tempdir);
 
 # Defining say here only due to wanting to be able to be run on perls
 # older than 5.10ish.
@@ -61,13 +63,41 @@ my $base_prefix    = ( defined $orb_base ) ? $orb_base : "$HOME/.orb";
 my $default_prefix = $base_prefix . q{/} . $lang_dir_map{$script_language};
 my $install_prefix = $default_prefix;
 my $cache_dir      = ( defined $orb_cache ) ? $orb_cache : "$base_prefix/cache";
-my $install_yaml   = 1;
-my $install_ffi    = 1;
-my $append_prefix  = 1;
-my $lang_vm        = 'default';
-my $lang_version   = 'default';
-my $rm_existing    = 0;
-my $suffix         = undef;
+my $log_dir        = ( defined $orb_logs ) ? $orb_logs : "$cache_dir\/logs";
+
+#say $base_prefix;
+#say $default_prefix;
+#say $install_prefix;
+#say $cache_dir;
+#say $log_dir;
+
+### no critic (ValuesAndExpressions::ProhibitMagicNumbers)
+#sleep 10;
+
+# handle orb_{ruby|perl|python}_base env variables.
+if ( $script_language eq 'ruby' ) {
+  $install_prefix =
+    ( defined $orb_ruby_base )
+    ? $orb_ruby_base
+    : $install_prefix;
+} elsif ( $script_language eq 'perl' ) {
+  $install_prefix =
+    ( defined $orb_perl_base ) ? $orb_perl_base : $install_prefix;
+} elsif ( $script_language eq 'python' ) {
+  $install_prefix =
+    ( defined $orb_python_base ) ? $orb_python_base : $install_prefix;
+}
+
+my $install_yaml  = 1;
+my $install_ffi   = 1;
+my $append_prefix = 1;
+my $lang_vm       = 'default';
+my $lang_version  = 'default';
+my $rm_existing   = 0;
+my $suffix        = undef;
+my $debug_flag    = 1;
+my $verbose_flag  = 1;
+my $tmp_base      = ( defined $TMPDIR ) ? $TMPDIR : '/tmp';
 
 # perl specific
 my $perl_install_cpanm = 0;
@@ -94,9 +124,9 @@ if ( $script_language eq 'ruby' ) {
   $help_message = <<FIN;
 usage:
 $0 --prefix=$install_prefix
-  --cache=$cache_dir
-  --ruby=[ruby|jruby|rbx] --version=[head|vm_version]
-  --no-yaml --no-ffi --rm
+   --cache=$cache_dir
+   --ruby=[ruby|jruby|rbx] --version=[head|vm_version]
+   --no-yaml --no-ffi --rm
 
   Options:
     --ruby     What ruby to install, ruby/rbx/jruby.
@@ -137,9 +167,15 @@ FIN
               'append'    => sub { $append_prefix = 1; },
               'rm'        => sub { $rm_existing = 1; },
               'suffix=s'  => \$suffix,
+              'tmp=s'     => \$tmp_base,
+              'debug'     => sub { $debug_flag = 0; },
+              'd'         => sub { $debug_flag = 0; },
+              'verbose'   => sub { $verbose_flag = 0; },
+              'v'         => sub { $verbose_flag = 0; },
               'help' => sub { help(1); },
               'h'    => sub { help(1); },
-            ) or help(1);
+            )
+    or help(1);
 
 } elsif ( $script_language eq 'perl' ) {
   $help_message = <<FIN;
@@ -179,14 +215,55 @@ FIN
     # yes this is wonky, its inverted for a later conditional
     # it makes the if test more readable... ish
     'no-test' => sub { $perl_run_tests = 0; },
-    'help'    => sub { help(1); },
-    'h'       => sub { help(1); },
-            ) or help(1);
+    'debug'   => sub { $debug_flag     = 0; },
+    'd'       => sub { $debug_flag     = 0; },
+    'tmp=s'   => \$tmp_base,
+    'verbose' => sub { $verbose_flag   = 0; },
+    'v'       => sub { $verbose_flag   = 0; },
+    'help' => sub { help(1); },
+    'h'    => sub { help(1); },
+            )
+    or help(1);
 
 } else {
   say $script_language;
   say "no idea what is being asked, was called as \"$script_name\"";
   exit 1;
+}
+
+# logs a commands output to:
+# $log_dir/$lang_vm-$lang_version.log
+sub log_cmd {
+  my $cmd = shift;
+  make_path $log_dir unless ( -d $log_dir );
+
+  my $logfile = "$log_dir\/$lang_vm-$lang_version\.log";
+  say $cmd;
+
+  my $full_cmd = undef;
+  if ( $debug_flag ^ $verbose_flag ) {
+    $full_cmd = "$cmd 2>&1 | tee -a $logfile";
+  } else {
+    $full_cmd = "$cmd 1>$logfile 2>&1";
+  }
+
+  system $full_cmd;
+
+  ## no critic (ValuesAndExpressions::ProhibitMagicNumbers)
+  my $rc = $? >> 8;    # should deal with signal/coredumps a better way
+
+  my $pwd = getcwd();
+
+  ## no critic (InputOutput::RequireCheckedOpen)
+  ## no critic (InputOutput::RequireCheckedClose)
+  ## no critic (InputOutput::RequireCheckedSyscalls)
+  ## no critic (InputOutput::RequireBracedFileHandleWithPrint)
+  open( my $fh, '>>', $logfile );
+  print $fh "$cmd return code was $rc\n";
+  print $fh "cwd was $pwd\n";
+  close $fh;
+
+  return $rc;
 }
 
 sub rubyurl {
@@ -201,7 +278,10 @@ sub perlurl {
   my $version = shift;
   $version =~ /(\d)[.]\d+/sm;
   ## no critic (RegularExpressions::ProhibitCaptureWithoutTest)
-  my $major   = $1 . '.0';
+  my $major =
+    ($1)
+    ? "$1\.0"
+    : die "Can't determine perl version from input $version\n";
   my $out_url = "http://www.cpan.org/src/$major/perl-$version.tar.gz";
   return $out_url;
 }
@@ -230,7 +310,7 @@ sub full_vm_version {
 
 sub sanity_check {
   my $insane = 0;
-  qx{ curl -V };
+  qx{curl -V};
   $insane = $insane ^ $?;
   warn 'Oh snap! Didn\'t find a working curl install.' if $?;
   qx{git --version};
@@ -241,6 +321,8 @@ sub sanity_check {
     make_path $cache_dir or die "Cannot create $cache_dir!\n";
   }
 
+  $File::Temp::DEBUG = !$debug_flag;
+
   full_vm_version();
 
   $append_prefix = 0 if ( $default_prefix ne $install_prefix );
@@ -250,9 +332,9 @@ sub sanity_check {
     if ($suffix) {
       $install_prefix = "$install_prefix" . "\%$suffix";
     }
-    if ( -d $install_prefix ) {
+    if ( -d $install_prefix and $rm_existing ) {
       say "Removing existing install $install_prefix";
-      system "rm -fr $install_prefix";
+      log_cmd "rm -fr $install_prefix";
     }
   }
 
@@ -262,11 +344,15 @@ sub sanity_check {
 
 sub curl_get {
   my $url = shift;
-  my $pwd = qx{pwd};
+  my $pwd = getcwd();
+  my $rc  = 0;
   chdir $cache_dir;
-  system "curl -L -O '$url'";
+
+  $rc = log_cmd "curl --progress-bar --location --remote-name '$url'";
+  die "curl failed to download $url\n" if $rc;
+
   chdir $pwd;
-  return $?;
+  return $rc;
 }
 
 sub url_filename {
@@ -282,11 +368,12 @@ sub fetch_url {
   my $url    = shift;
   my $exists = "$cache_dir/" . url_filename($url);
   my $ok     = 0;
+  chdir $cache_dir;
+
   if ( -f $exists ) {
     say "Found $exists will not redownload.";
   } else {
     say "$exists not found\nTrying to download $url";
-
     my $rc = curl_get($url);
     warn "nonzero exit from curl $rc on url $url" if $rc;
     $ok = $ok ^ $rc;
@@ -308,50 +395,59 @@ sub full_rubyversion {
 }
 
 sub extract_tgz {
-  my $file = shift;
-  my $pwd  = qx{pwd};
-  say "gunzip -c $file | tar xf -";
-  chdir $cache_dir;
-  system "gunzip -c $file | tar xf -";
-  die "Extract of $file failed.\n" if $?;
-  chdir $pwd;
-  return $?;
+  my $file    = shift;
+  my $tmp_dir = tempdir( DIR => $tmp_base, CLEANUP => 1 );
+  my $rc      = 0;
+
+  chdir $tmp_dir;
+
+  $rc = log_cmd "gunzip -c $file | tar xf -";
+  die "Extract of $file failed.\n" if $rc;
+  return $tmp_dir;
 }
 
 sub configure_dir {
   my $dir  = shift;
   my $args = shift;
+  my $rc   = 0;
   chdir $dir;
-  system "./configure $args";
-  die "Configure failed\n" if $?;
-  return $?;
+
+  $rc = log_cmd "./configure $args";
+  die "Configure failed\n" if $rc;
+  return;
 }
 
-sub run_make {
+sub build_with {
+  my $tool = shift;
   my $dir  = shift;
   my @rest = @_;
-  say $dir;
-
+  my $rc   = 0;
   chdir $dir;
-  system 'make';
-  die "make failed $?\n" if $?;
-  for my $cmd (@rest) {
-    my $make_cmd = "make $cmd";
-    system $make_cmd;
-    die "$make_cmd failed\n" if $?;
+
+  # default rake task includes testing, NOPE
+  unless ( $lang_vm eq 'rbx' ) {
+    $rc = log_cmd $tool;
+    die "$tool failed $rc\n" if $rc;
   }
-  return $?;
+
+  for my $cmd (@rest) {
+    my $make_cmd = "$tool $cmd";
+    $rc = log_cmd $make_cmd;
+    die "$make_cmd failed\n" if $rc;
+  }
+  return $rc;
 }
 
 sub install_yaml {
-  my $url      = 'http://pyyaml.org/download/libyaml/yaml-0.1.4.tar.gz';
-  my $dirname  = 'yaml-0.1.4';
-  my $filename = "$cache_dir/" . url_filename($url);
-  chdir $cache_dir;
+  say 'Installing libyaml.';
+  my $yaml_version = '0.1.4';
+  my $destdir      = "yaml-$yaml_version";
+  my $url          = "http://pyyaml.org/download/libyaml/$destdir.tar.gz";
+  my $filename     = url_filename($url);
   fetch_url($url);
-  extract_tgz($filename);
-  configure_dir( $dirname, "--prefix=$install_prefix" );
-  run_make( $dirname, 'install' );
+  my $dir = extract_tgz("$cache_dir\/$filename") . q{/} . $destdir;
+  configure_dir( $dir, "--prefix=$install_prefix" );
+  build_with( 'make', $dir, 'install' );
   return;
 }
 
@@ -361,154 +457,191 @@ sub fetch_git {
   my $branch   = shift || 'master';
   my $full_dir = "$cache_dir\/$dirname";
 
+  my $rc = 0;
   if ( -d $full_dir ) {
     chdir $full_dir;
-    system 'git pull';
-    die "git pull of $url failed\n" if $?;
-    system 'git reset --hard HEAD';
-    die "git reset hard failed\n" if $?;
+    $rc = log_cmd 'git pull';
+    die "git pull of $url failed\n" if $rc;
+    $rc = log_cmd 'git reset --hard HEAD';
+    die "git reset hard failed\n" if $rc;
   } else {
     my $cmd = "git clone $url $full_dir";
     $cmd = "git clone --branch $branch $url $full_dir"
       unless $branch eq 'master';
-    system $cmd;
-    die "git clone of $url failed\n" if $?;
+    $rc = log_cmd $cmd;
+    die "git clone of $url failed\n" if $rc;
   }
-  return;
+
+  # k so now shallow clone to a temporary directory
+  my $tmp_dir = tempdir( DIR => $tmp_base, CLEANUP => 1 );
+  my $cmd = "git clone --depth 1 file://$full_dir $tmp_dir";
+  $rc = log_cmd $cmd;
+  die "git shallow clone failed\n" if $rc;
+  return $tmp_dir;
 }
 
 sub install_ffi {
+  say 'Installing ffi.';
   my $dirname = 'libffi-git';
-  fetch_git( 'https://github.com/atgreen/libffi', $dirname );
-  chdir "$cache_dir/$dirname";
-  system 'make clean';
-  configure_dir( $dirname, "--prefix=$install_prefix" );
-  run_make( $dirname, 'install' );
+  my $dir = fetch_git( 'https://github.com/atgreen/libffi', $dirname );
+  configure_dir( $dir, "--prefix=$install_prefix" );
+  build_with( 'make', $dir, 'install' );
   return;
 }
 
 sub install_ruby {
-  $LDFLAGS = ( defined($LDFLAGS) ? $LDFLAGS : q{} ) . " -L$install_prefix/lib";
-  $CPPFLAGS =
-    ( defined($CPPFLAGS) ? $CPPFLAGS : q{} ) . " -L$install_prefix/include";
+  if ( $install_ffi or $install_yaml ) {
+    if ( defined($LDFLAGS) ) {
+      $LDFLAGS = "$LDFLAGS -L$install_prefix/lib";
+    }
 
-  install_ffi  if $install_ffi;
-  install_yaml if $install_yaml;
+    if ( defined($CPPFLAGS) ) {
+      $CPPFLAGS = "$CPPFLAGS -L$install_prefix/include";
+    }
 
-  my $dirname = q{};
+    install_ffi()  if $install_ffi;
+    install_yaml() if $install_yaml;
+    tmp_cleanup();
+  }
+
+  my $dir = q{};
   my $configure =
 "--prefix=$install_prefix --disable-tk --disable-tcl --disable-install-doc --enable-pthread --with-opt-dir=$install_prefix --with-static-linked-ext --disable-tcltk";
+
   if ( $lang_version ne 'head' ) {
     my $url  = rubyurl("$lang_version");
     my $file = url_filename($url);
-    $dirname = "$cache_dir/$file";
-    $dirname =~ s/[.]tar[.]gz//sm;
+
+    $dir = $file;
+    $dir =~ s/[.]tar[.]gz//sm;
+
     my $filename = "$cache_dir/$file";
+
     fetch_url($url);
-    extract_tgz($filename);
-    chdir "$cache_dir\/$dirname";
-    configure_dir( $dirname, $configure );
+
+    $dir = extract_tgz($filename) . q{/} . $dir;
   } else {
-    $dirname = 'ruby-git';
-    fetch_git( 'https://github.com/ruby/ruby', $dirname );
-    chdir "$cache_dir\/$dirname";
-    system 'make clean';
-    system 'autoconf';
-    if ( $^O eq 'darwin' ) {
-      $configure = "CC=clang $configure";
-    }
-    configure_dir( $dirname, "$configure" );
+    $dir = 'ruby-git';
+    $dir = fetch_git( 'https://github.com/ruby/ruby', $dir );
+
+    chdir $dir;
+    log_cmd 'autoconf';
   }
-  run_make( $dirname, 'install' );
+
+  if ( $^O eq 'darwin' ) {
+    $configure = "CC='clang' $configure";
+
+    #      $configure = "CC='clang -Qunused-arguments' $configure";
+  }
+
+  configure_dir( $dir, "$configure" );
+
+  build_with( 'make', $dir, 'install' );
+
   return;
 }
 
 sub install_rbx {
-  $LDFLAGS = ( defined($LDFLAGS) ? $LDFLAGS : q{} ) . " -L$install_prefix/lib";
-  $CPPFLAGS =
-    ( defined($CPPFLAGS) ? $CPPFLAGS : q{} ) . " -L$install_prefix/include";
+  if ($install_yaml) {
+    if ( defined($LDFLAGS) ) {
+      $LDFLAGS = "$LDFLAGS -L$install_prefix/lib";
+    }
 
-  install_yaml if $install_yaml;
+    if ( defined($CPPFLAGS) ) {
+      $CPPFLAGS = "$CPPFLAGS -L$install_prefix/include";
+    }
+
+    install_yaml();
+    tmp_cleanup();
+  }
 
   chdir $cache_dir;
+
   my $dirname = 'rbx-git';
   my $fulldir = "$cache_dir\/$dirname";
   my $configure =
 "--prefix=$install_prefix --default-version=19 --with-include-dir=$install_prefix/include --with-lib-dir=$install_prefix/lib";
-  fetch_git( 'https://github.com/rubinius/rubinius', $dirname );
-  chdir $fulldir;
-  system 'rake clean';
-  configure_dir( $dirname, $configure );
-  system 'rake install';
-  return;
-}
 
-sub install_jruby {
-  chdir $cache_dir;
-  my $dirname = q{};
-  if ( $lang_version ne 'head' ) {
-    my $url =
-"http://jruby.org.s3.amazonaws.com/downloads/$lang_version/jruby-bin-$lang_version.tar.gz";
-    my $file = url_filename($url);
-    $dirname = $file;
-    $dirname =~ s/[.]tar[.]gz//sm;
-    $dirname =~ s/bin[-]//sm;
+  my $dir = fetch_git( 'https://github.com/rubinius/rubinius', $dirname );
+  my $prebuilt_dir = "$fulldir\/vendor\/prebuilt";
 
-    my $extract_dir = $dirname;
-    $extract_dir =~ s/[-]bin//sm;
-    my $filename = "$cache_dir/$file";
-    fetch_url($url);
-    extract_tgz($filename);
+  if ( -d $prebuilt_dir ) {
+    say 'Copying cache prebuilt llvm files to the temporary git clone.';
+    chdir "$fulldir\/vendor";
+    log_cmd "find ./prebuilt -print -depth | cpio -pdmv $dir\/vendor";
+  }
 
-    say 'Removing windows stuffs.';
-    my $cmd =
-"rm -f $extract_dir/bin/*.exe $extract_dir/bin/*.dll $extract_dir/bin/*.bat $extract_dir/bin/jruby.sh";
-    say $cmd;
-    system $cmd;
-    system "mv $extract_dir $install_prefix";
-    chdir "$install_prefix\/bin";
-    system 'ln -s jruby ruby';
-  } else {
-    $dirname = 'jruby-git';
-    fetch_git( 'https://github.com/jruby/jruby', $dirname );
-    chdir "$cache_dir\/$dirname";
-    chdir 'jruby';
-    system 'ant clean';
-    system 'ant';
-    system 'ant dist';
-    chdir 'dist';
-    my $dir = qx(pwd);
-    chomp($dir);
-    my $distfile      = glob('jruby-bin*tar.gz');
-    my $distfile_full = "$dir/" . $distfile;
+  chdir $dir;
 
-    if ( -f $distfile ) {
-      extract_tgz($distfile_full);
-      my $extract_dir = $distfile;
-      $extract_dir =~ s/[.]tar[.]gz//sm;
-      $extract_dir =~ s/[-]bin//sm;
-      say $extract_dir;
-      system "mv $extract_dir $install_prefix";
-      jruby_finalize($install_prefix);
-    } else {
-      say 'zomg';
-    }
+  configure_dir( $dir, $configure );
+  build_with( 'rake', $dir, 'install' );
+
+  # Be kind bandwidth wise, copy whatevers in the temp prebuilt dir
+  # back to the cache prebuilt dir for later use.
+  $prebuilt_dir = "$dir\/vendor\/prebuilt";
+  if ( -d $prebuilt_dir ) {
+    say 'Copying temporary prebuilt llvm files back to caches.';
+    chdir "$dir\/vendor";
+    log_cmd "find ./prebuilt -print -depth | cpio -pdmv $fulldir\/vendor";
   }
   return;
 }
 
+sub install_jruby {
+  my $dir = q{};
+
+  if ( $lang_version ne 'head' ) {
+    my $url =
+"http://jruby.org.s3.amazonaws.com/downloads/$lang_version/jruby-bin-$lang_version.tar.gz";
+    my $file = url_filename($url);
+    $dir = $file;
+    $dir =~ s/[.]tar[.]gz//sm;
+    $dir =~ s/bin[-]//sm;
+
+    my $extract_dir = $dir;
+    $extract_dir =~ s/[-]bin//sm;
+    my $filename = "$cache_dir/$file";
+    fetch_url($url);
+    $dir = extract_tgz("$cache_dir\/$filename");
+
+    log_cmd "mv $extract_dir $install_prefix";
+  } else {
+    $dir = 'jruby-git';
+    $dir = fetch_git( 'https://github.com/jruby/jruby', $dir );
+    build_with( 'ant', $dir, 'dist' );
+    chdir "$dir\/dist";
+
+    my $distfile      = glob('jruby-bin*tar.gz');
+    my $distfile_full = "$dir/dist/" . $distfile;
+
+    if ( -f $distfile ) {
+      $dir = extract_tgz($distfile_full);
+      my $extract_dir = $distfile;
+      $extract_dir =~ s/[.]tar[.]gz//sm;
+      $extract_dir =~ s/[-]bin//sm;
+
+      log_cmd "mv $extract_dir $install_prefix";
+    } else {
+      say 'Jruby build failed for some reason or another.';
+    }
+  }
+  jruby_finalize($install_prefix);
+  return;
+}
+
 sub jruby_finalize {
-  my $pwd = qx(pwd);
   my $dir = shift;
-  say 'Removing windows stuffs.';
+
+  say 'Removing windows specific files.';
+
   my $cmd =
     "rm -f $dir/bin/*.exe $dir/bin/*.dll $dir/bin/*.bat $dir/bin/jruby.sh";
-  say $cmd;
-  system $cmd;
+
+  log_cmd $cmd;
   my $ruby_link = "$dir\/bin/ruby";
   chdir "$dir\/bin";
-  system 'ln -s jruby ruby' unless ( -f $ruby_link );
-  chdir $pwd;
+  log_cmd 'ln -s jruby ruby' unless ( -f $ruby_link );
+
   return;
 }
 
@@ -521,16 +654,13 @@ sub perl_finalize {
     unless ($perl_install_cpanm) {
       say "Setting up cpanm for $perl_install";
       my $cpanm_install_cmd =
-        "curl -L http://cpanmin.us | $perl_bin - --self-upgrade";
-      say $cpanm_install_cmd;
-      system($cpanm_install_cmd);
+        "curl --silent --location http://cpanmin.us | $perl_bin -";
+      log_cmd $cpanm_install_cmd;
     }
   }
 }
 
 sub install_perl {
-  my $pwd = getcwd();
-
   my $dirname = q{};
   my $configure =
 "sh Configure -ds -e -Dprefix=$install_prefix -Dvendorprefix=$install_prefix -D optimize='-Os -Wall -g -pipe' -U use5005threads -D usethreads -D usemultiplicity -D uselargefiles -D use64bitint -D useshrplib";
@@ -542,27 +672,29 @@ sub install_perl {
   if ( $lang_version ne 'head' ) {
     my $url  = perlurl($lang_version);
     my $file = url_filename($url);
-    $dirname = "$cache_dir/$file";
+    $dirname = $file;
     $dirname =~ s/[.]tar[.]gz//sm;
     my $filename = "$cache_dir/$file";
     fetch_url($url);
-    extract_tgz($filename);
-    chdir $dirname;
+    my $dir = extract_tgz($filename) . q{/} . $dirname;
 
-    # just in case, don't want any cached values being used.
-    unlink('config.sh') if ( -f 'config.sh' );
-    system($configure);
+    chdir $dir;
+    log_cmd $configure;
 
     if ($perl_run_tests) {
-      run_make( $dirname, 'test', 'install' );
+      build_with( 'make', $dir, 'test', 'install' );
     } else {
-      say 'Not testing installation prior to install';
-      run_make( $dirname, 'install' );
+      build_with( 'make', $dir, 'install' );
     }
   }
 
   perl_finalize($install_prefix);
   return;
+}
+
+sub tmp_cleanup {
+  chdir q{/};
+  File::Temp::cleanup();
 }
 
 my %install_switch = (
@@ -578,3 +710,7 @@ if ( sanity_check() ) {
 } else {
   $install_switch{$lang_vm}();
 }
+
+local $SIG{'INT'} = sub { tmp_cleanup(); };
+
+END { tmp_cleanup(); }
